base numpy window implementation & sma
