+++ temporian/implementation/numpy/operators/all_operators.py
@@ -6,7 +6,9 @@ from temporian.implementation.numpy.operators import assign
 from temporian.implementation.numpy.operators import lag
 from temporian.implementation.numpy.operators import propagate
 from temporian.implementation.numpy.operators import select
-from temporian.implementation.numpy.operators import simple_moving_average
+from temporian.implementation.numpy.operators.window import (
+    simple_moving_average,
+)
 from temporian.implementation.numpy.operators.calendar import day_of_month
 from temporian.implementation.numpy.operators.calendar import day_of_week
